Render projects server side
