+++ src/app/projects/page.tsx
@@ -1,10 +1,11 @@
-import PageWrapper from '@/components/core/page/PageWrapper';
-import ListProjects from '@/components/projects/ListProjects';
-import { ProjectTypes } from '@/types/deprecated/models/enums';
+import ListProjects from '@/components/projects/List/ListProjects';
+import { ProjectFragment } from '@/graphql/generated/generated-types';
+import getApiResponse from '@/utils/api/getApiResponse';
 import { getSpaceServerSide } from '@/utils/api/getSpaceServerSide';
 
 export default async function ProjectListPage() {
   const space = await getSpaceServerSide();
+  const projects = await getApiResponse<ProjectFragment[]>(space!, 'projects');
 
-  return <ListProjects space={space!} type={'All'} />;
+  return <ListProjects space={space!} projects={projects} />;
 }
