+++ src/app/projects/type/[projectType]/page.tsx
@@ -1,6 +1,8 @@
 import PageWrapper from '@/components/core/page/PageWrapper';
-import ListProjects from '@/components/projects/ListProjects';
+import ListProjects from '@/components/projects/List/ListProjects';
+import { ProjectFragment } from '@/graphql/generated/generated-types';
 import { ProjectTypes } from '@/types/deprecated/models/enums';
+import getApiResponse from '@/utils/api/getApiResponse';
 import { getSpaceServerSide } from '@/utils/api/getSpaceServerSide';
 
 export default async function ProjectListPage(props: {
@@ -11,9 +13,11 @@ export default async function ProjectListPage(props: {
   const space = await getSpaceServerSide();
   const type = Object.keys(ProjectTypes).find((key) => key.toLowerCase() === props.params.projectType.toLowerCase());
 
+  const projects = await getApiResponse<ProjectFragment[]>(space!, 'projects');
+
   return (
     <PageWrapper>
-      <ListProjects space={space!} type={type!} />
+      <ListProjects space={space!} projects={projects} />
     </PageWrapper>
   );
 }
