+++ src/app/projects/view/[projectId]/[viewType]/layout.tsx
@@ -1,15 +1,17 @@
 import PageWrapper from '@/components/core/page/PageWrapper';
 import { ViewProjectHeader } from '@/components/projects/View/ViewProjectHeader';
-import { getProjectUsingApi } from '@/utils/api/project/getProjectUsingApi';
+import { ProjectFragment } from '@/graphql/generated/generated-types';
+import getApiResponse from '@/utils/api/getApiResponse';
 import { getSpaceServerSide } from '@/utils/api/getSpaceServerSide';
 import React from 'react';
 
 async function ProjectViewHome(props: { params: { projectId: string; viewType: string }; children: React.ReactNode }) {
   const space = await getSpaceServerSide();
-  const project = await getProjectUsingApi(props.params.projectId);
+  const project = await getApiResponse<ProjectFragment>(space!, `projects/${props.params.projectId}`);
+
   return (
     <PageWrapper>
-      <ViewProjectHeader project={project!} selectedViewType={props.params.viewType} />
+      <ViewProjectHeader project={project} selectedViewType={props.params.viewType} />
       <div className="mt-8">{props.children}</div>
     </PageWrapper>
   );
