+++ src/components/projects/List/ProjectSummaryCard.tsx
@@ -1,0 +1,35 @@
+import styles from '@/components/projects/List/ListProjectsHelper.module.scss';
+import ProjectSummaryCardAdminDropdown from '@/components/projects/List/ProjectSummaryCardAdminDropdown';
+import { ProjectFragment, SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
+import Link from 'next/link';
+import React from 'react';
+
+export interface ProjectSummaryCardProps {
+  space: SpaceWithIntegrationsFragment;
+  project: ProjectFragment;
+}
+
+export default function ProjectSummaryCard({ space, project }: ProjectSummaryCardProps) {
+  return (
+    <div key={project.id} className="rounded-lg overflow-hidden shadow-xl">
+      <Link href={`/projects/view/${project.id}/tidbit-collections`}>
+        <div className={`h-[250px] ${styles.card} px-1.5 pt-2 rounded-lg flex flex-col gap-2.5 shadow-xl`}>
+          <div className="rounded relative">
+            <img
+              src={project.cardThumbnail ?? 'https://picsum.photos/500/300'}
+              alt={project.name}
+              className="h-[150px] w-full object-cover rounded shadow-[0_10px_13px_-5px_rgba(0,0,0,0.6)]"
+            />
+            <div className="absolute top-0 right-0 m-2">
+              <ProjectSummaryCardAdminDropdown space={space} project={project} />
+            </div>
+          </div>
+          <div className={`${styles.header} mx-2`}>
+            <div className="font-bold">{project.name}</div>
+            <div className={`text-sm ${styles.details}`}>{project.excerpt}</div>
+          </div>
+        </div>
+      </Link>
+    </div>
+  );
+}
