+++ src/components/projects/List/ProjectSummaryCardAdminDropdown.tsx
@@ -1,0 +1,92 @@
+'use client';
+
+import DeleteConfirmationModal from '@/components/app/Modal/DeleteConfirmationModal';
+import PrivateEllipsisDropdown from '@/components/core/dropdowns/PrivateEllipsisDropdown';
+import UpsertProjectModal from '@/components/projects/Edit/UpsertProjectModal';
+import { useNotificationContext } from '@/contexts/NotificationContext';
+import { ProjectFragment, SpaceWithIntegrationsFragment, useUpdateArchivedStatusOfProjectMutation } from '@/graphql/generated/generated-types';
+import React, { useState } from 'react';
+
+export interface ProjectSummaryCardProps {
+  space: SpaceWithIntegrationsFragment;
+  project: ProjectFragment;
+}
+
+export default function ProjectSummaryCardAdminDropdown({ space, project }: ProjectSummaryCardProps) {
+  const [editProject, setEditProject] = useState<ProjectFragment | null>(null);
+  const [deleteProjectId, setDeleteProjectId] = useState<string | null>(null);
+  const [updateArchivedStatusOfProjectMutation] = useUpdateArchivedStatusOfProjectMutation();
+  const { showNotification } = useNotificationContext();
+
+  const getThreeDotItems = (project: ProjectFragment) => {
+    if (project.archived) {
+      return [
+        { label: 'Edit', key: 'edit' },
+        { label: 'Unarchive', key: 'unarchive' },
+      ];
+    }
+    return [
+      { label: 'Edit', key: 'edit' },
+      { label: 'Archive', key: 'archive' },
+    ];
+  };
+
+  const updateArchiveStatus = async (projectId: string, archived: boolean) => {
+    try {
+      await updateArchivedStatusOfProjectMutation({
+        variables: {
+          projectId: projectId,
+          archived: archived,
+        },
+        refetchQueries: ['Projects'],
+      });
+      if (archived) {
+        showNotification({ message: 'Project archived successfully', type: 'success' });
+      } else {
+        showNotification({ message: 'Project un-archived successfully', type: 'success' });
+      }
+    } catch (error) {
+      showNotification({ message: 'Something went wrong', type: 'error' });
+    }
+  };
+
+  return (
+    <>
+      <PrivateEllipsisDropdown
+        items={getThreeDotItems(project)}
+        onSelect={async (key) => {
+          if (key === 'edit') {
+            setEditProject(project);
+          } else if (key === 'archive') {
+            setDeleteProjectId(project.id);
+            setDeleteProjectId(project.id);
+          } else if (key === 'unarchive') {
+            console.log('unarchive');
+            updateArchiveStatus(project.id, false);
+          }
+        }}
+      />
+      {deleteProjectId && (
+        <DeleteConfirmationModal
+          title={'Delete Project'}
+          open={!!deleteProjectId}
+          onClose={() => setDeleteProjectId(null)}
+          onDelete={() => {
+            updateArchiveStatus(deleteProjectId, true);
+            setDeleteProjectId(null);
+          }}
+        />
+      )}
+      {editProject && (
+        <UpsertProjectModal
+          spaceId={space.id}
+          project={project || undefined}
+          open={!!editProject}
+          onClose={() => {
+            setEditProject(null);
+          }}
+        />
+      )}
+    </>
+  );
+}
