+++ src/components/spaces/manageSpaceSubviews.ts
@@ -1,7 +1,5 @@
 export enum ManageSpaceSubviews {
   SpacesList = 'spaces',
-  ProjectList = 'projects',
-  ProjectDetails = 'project-details',
   Chatbot = 'chatbot',
   ViewSpace = 'view-space',
   GenerateImage = 'generate-image',
